Create Marketplace.sol
Three states variables: buyer, seller, and balance
Two events: list Item and item Purchased
One enum with two values: item available and item purchased
Consumer assigns the seller user as msg.sender and sets the initial state to item available. This constructor is called when the contract is created. The buyer function tales three parameters: seller, buyer, and price. It has a requirement that the buyer has enough money for the purchase, then it transfers money from the buyer to the seller, and finally, a message is emitted.